Simplify ObjectBVH construction
